refactoring: clause managers are packed into ClauseManager module
